v1.4(mock):
add json mock data parser (at app start), minor updates

+++ src/main/java/aviatickets/app/flight/dto/request/RequestFlightItem.java
@@ -16,7 +16,7 @@ public record RequestFlightItem(
 		List<Leg> itinerary,
 		Aircraft aircraft,
 
-		Integer totalDistance, // => leg distance += leg distance
+		Short totalDistance, // => leg distance += leg distance
 		String totalDuration,
 		Price price,
 		Short passengerCount,
